Match OLMo-core SFT experiment: OBFD packing + numpy RNG ordering Co-Authored-By: Claude Opus 4.6 <noreply@anthropic.com>
